Migrate TorLauncher, ads, ledger services off service manager

+++ components/services/tor/tor_launcher_impl.cc
@@ -76,14 +76,19 @@ static void TearDownPipeHack() {
 namespace tor {
 
 TorLauncherImpl::TorLauncherImpl(
-    std::unique_ptr<service_manager::ServiceContextRef> service_ref)
-    : service_ref_(std::move(service_ref)) {
+    mojo::PendingReceiver<mojom::TorLauncher> receiver)
+    : receiver_(this, std::move(receiver)) {
+  receiver_.set_disconnect_handler(
+      base::BindOnce(&TorLauncherImpl::Cleanup, base::Unretained(this)));
 #if defined(OS_POSIX)
   SetupPipeHack();
 #endif
 }
 
-TorLauncherImpl::~TorLauncherImpl() {
+void TorLauncherImpl::Cleanup() {
+  if (in_shutdown_) return;
+  in_shutdown_ = true;
+
   if (tor_process_.IsValid()) {
     tor_process_.Terminate(0, true);
 #if defined(OS_POSIX)
@@ -98,6 +103,14 @@ TorLauncherImpl::~TorLauncherImpl() {
     base::EnsureProcessTerminated(std::move(tor_process_));
 #endif
   }
+}
+
+TorLauncherImpl::~TorLauncherImpl() {
+  Cleanup();
+}
+
+void TorLauncherImpl::Shutdown() {
+  Cleanup();
 }
 
 void TorLauncherImpl::Launch(const TorConfig& config,
@@ -198,7 +211,7 @@ void TorLauncherImpl::MonitorChild() {
             LOG(ERROR) << "tor exit (" << WEXITSTATUS(status) << ")";
           }
           tor_process_.Close();
-          if (connected_ && crash_handler_callback_) {
+          if (receiver_.is_bound() && crash_handler_callback_) {
             base::ThreadTaskRunnerHandle::Get()->PostTask(
               FROM_HERE, base::BindOnce(std::move(crash_handler_callback_),
                                         pid));
@@ -211,17 +224,13 @@ void TorLauncherImpl::MonitorChild() {
   }
 #elif defined(OS_WIN)
   WaitForSingleObject(tor_process_.Handle(), INFINITE);
-  if (connected_ && crash_handler_callback_)
+  if (receiver_.is_bound() && crash_handler_callback_)
     base::ThreadTaskRunnerHandle::Get()->PostTask(
       FROM_HERE, base::BindOnce(std::move(crash_handler_callback_),
                                 base::GetProcId(tor_process_.Handle())));
 #else
 #error unsupported platforms
 #endif
-}
-
-void TorLauncherImpl::SetDisconnected() {
-  connected_ = false;
 }
 
 }  // namespace tor
